Move capabilitiesEditable to capabilities folder
